requirements.txt
alembic==1.17.1
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.11.0
awscli==1.42.72
bcrypt==4.0.1
boto3==1.40.71
botocore==1.40.72
certifi==2025.11.12
click==8.3.0
colorama==0.4.6
coverage==7.11.3
debugpy==1.8.17
dnspython==2.8.0
docutils==0.19
ecdsa==0.19.1
email-validator==2.3.0
exceptiongroup==1.3.0
fastapi==0.121.1
greenlet==3.2.4
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
iniconfig==2.3.0
jmespath==1.0.1
Mako==1.3.10
MarkupSafe==3.0.3
mediatr==1.3.2
packaging==25.0
passlib==1.7.4
pluggy==1.6.0
psycopg2==2.9.11
pyasn1==0.6.1
pydantic==2.12.4
pydantic-settings==2.12.0
pydantic_core==2.41.5
Pygments==2.19.2
pytest==9.0.0
pytest-cov==7.0.0
pytest-dotenv==0.5.2
pytest-mock==3.15.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-jose==3.5.0
python-multipart==0.0.20
PyYAML==6.0.3
rsa==4.7.2
s3transfer==0.14.0
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.44
starlette==0.49.3
tomli==2.3.0
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.5.0
uvicorn==0.38.0
